// src/app/model/User.ts
export class User {
    id: number;
    role_id: number;
    account_enabled: Boolean;
    name: String ;
    full_name: String;
    password: String;
    confirm_password: String;
    email_id: String;
    address: String;
    contact_number: String;
    imagePath:string;
    libraryName:string;
}